CreditCard records are marked deleted, but never destroyed
Charge records can reference a CreditCard, so we need them to persist.

Before destroying a CreditCard, we overwrite the number with a fake
number (41111) that still passes the Luhn check.  This is the ensure
that if you tell a customer, for example, that their credit card
information has been deleted, then there is no way you could charge the
credit card again.