Prefixed the Sellable module instance methods with "sellable_"
Turns out that the names "code", "description", and "amount" are very
common and can easily conflict with names of attributes that one may
already have present in a model but used for a different purpose.  I
originally picked these common names because I assumed they would be
used for the *same* purpose, but I can't be sure of that.

Therefore, we won't clobber those methods and instead namespace our own.